Mark "unconfirmed" businesses on map in place of MCX.
Merchant Customer Exchange (MCX) members were originally tracked on the
map due to a boycott following the release of Apple Pay and the
disablement of NFC at CVS and Rite-Aid. Due to former MCX members
jumping off the sinking ship and a realization that such tracking is out
of scope for this site, this functionality is now removed.

In its place is the notion of "unconfirmed" businesses. These are
locations that are part of a larger chain that is known to be rolling
out EMV but have not been confirmed to have that functionality. This
allows businesses like Home Depot (which is currently rolling out EMV)
to be represented on the map even though there is no confirmation from
their PR as to the current extent. Users can confirm these locations by
reporting them as appropriate.