Major change: allow multiple businesses in a single location.
In the previous code, there was a 1:1 mapping between pins and businesses. This
could cause trouble when adding certain locations inside institutions such as
shopping malls. Google autocorrect helps with that somewhat by knowing the
locations of some businesses within major malls but there are still gaps. Changes
in this changelist are threefold:

1. A slight change to the SQL queries sent by the Scala backend to ensure that
   for each unique random location retrieved, all businesses at that location
   are sent to the client.
2. Yellow pins now appear on the map when there is more than one business at
   a specific latitude/longitude pair.
3. GUI updates to be able to show each business at a specific location. This
   mainly revolves around rendering a list of all businesses below the normal
   "add business" popup.

"Unseen" changes include a change from Mustache to Handlebars for JS templating
and a significant refactoring of main.coffee to accomodate the above changes
and hopefully improve performance. I tested this a fair amount and there are
probably still issues but hopefully they aren't major.